Setup inotify on the parent directory and filter events

+++ watcher.go
@@ -17,6 +17,7 @@
 package main
 
 import (
+	"path"
 	"syscall"
 
 	"gopkg.in/fsnotify.v1"
@@ -27,14 +28,19 @@ func watch(files []string) {
 	panicOnError(err)
 
 	for _, file := range files {
-		watcher.Add(file)
+		watcher.Add(path.Dir(file))
 	}
 
 	for {
 		select {
 		case event := <-watcher.Events:
-			logger.Printf("found new %s, reloading", event.Name)
-			syscall.Kill(syscall.Getpid(), syscall.SIGUSR1)
+			for _, file := range files {
+				if event.Name == path.Base(file) {
+					logger.Printf("detected change on %s, reloading", event.Name)
+					syscall.Kill(syscall.Getpid(), syscall.SIGUSR1)
+					break
+				}
+			}
 
 		case err := <-watcher.Errors:
 			logger.Printf("error watching file: %s", err)
